Clean up handling of file overwrites
